+++ src/commands/history.js
@@ -1,63 +1,12 @@
 const chalk = require("chalk");
 
+const _addToHist = require("../functions/addToHist");
 const _fatalError = require("../functions/fatalError");
 
 const Errors = require("../classes/Errors");
 const InfoMessages = require("../classes/InfoMessages");
 const ConfigManager = require("../classes/ConfigManager");
 const Verbose = require("../classes/Verbose");
-
-/**
- * The number of history commands to store before deleting the oldest ones.
- */
-const NUMBER_TO_STORE = 50;
-
-/**
- * Add a command to the history to be later
- * shown in the BubbleOS command `history`.
- *
- * This command appends to an array, which
- * is stored in the JavaScript heap. This means
- * that when BubbleOS is shut down, the history
- * will be cleared.
- *
- * @param {string} command The command that the user entered that should be stored in the history.
- * @param {boolean} addToConfig Whether or not to add the command to the BubbleOS configuration. Defaults to `true`.
- */
-const _addToHist = (command, addToConfig = true) => {
-  if (!addToConfig) return;
-
-  const config = new ConfigManager();
-
-  // Fetch the history from the config
-  Verbose.custom("Checking if configuration exists...");
-  if (typeof config.getConfig() === "undefined") {
-    InfoMessages.error(
-      "Error when saving command to history in the configuration file. Resetting file..."
-    );
-
-    Verbose.custom("Resetting configuration file...");
-    config.deleteConfig();
-    config.createConfig();
-    return;
-  }
-
-  Verbose.custom("Getting history stored in configuration file...");
-  const historyConfig = config.getConfig().history ?? [];
-
-  // If the number of stored commands exceeds the limit, remove the oldest entry
-  Verbose.custom(
-    `Checking if number of entries stored is greater than ${NUMBER_TO_STORE} entries, and deleting oldest if so...`
-  );
-  if (historyConfig.length + 1 > NUMBER_TO_STORE) historyConfig.shift();
-
-  // Add the latest command to the history
-  Verbose.custom("Adding latest command to history...");
-  historyConfig.push(command);
-
-  Verbose.custom("Writing data to configuration file...");
-  config.addData({ history: historyConfig });
-};
 
 /**
  * Formats history index and command, and logs it.
@@ -172,4 +121,4 @@ const history = (numToDisplay, ...args) => {
   }
 };
 
-module.exports = { history, _addToHist };
+module.exports = history;